requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.2
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.4
python-multipart==0.0.9
pytest==8.0.0
httpx==0.26.0 